backend/requirements.txt
fastapi==0.109.0
fastapi-cli
openai
python-dotenv